cmd/ci-operator: Add --grace-period
Because while two seconds may be sufficient teardown time for unit
tests and such, it's too short for e2e teardown (where cluster logs
should be collected, cluster resources need to be reaped, and
collected assets need to be uploaded).  The two-second timeout is from
c2ac3abf (When interrupted, mark any in progress pods/templates as
deleted, 2018-05-04, #16).  And actually, I'm not sure how the
2-second default worked even for unit tests and such, because
Kubernetes pods have a 30-second grace period by default [1].

A recent aborted e2e job that did not run teardown to completion or
collect any cluster assets is [2,3].  A recent successful e2e job
showing current teardown timing is [4,5].  From [4]:

  2019/02/14 18:19:54 Container setup in pod e2e-aws completed successfully
  2019/02/14 18:46:08 Container test in pod e2e-aws completed successfully
  2019/02/14 18:51:38 Container teardown in pod e2e-aws completed successfully

So 5.5 minutes to teardown.  And from [5]:

  time="2019-02-14T18:47:40Z" level=debug msg="search for and delete matching resources by tag in us-east-1 matching aws.Filter{\"openshiftClusterID\":\"1341ccc2-66ac-46bb-ae15-0295d4a126ba\"}"
  ...
  time="2019-02-14T18:51:38Z" level=debug msg="Purging asset \"Cluster\" from disk"

So about 4 minutes of that is resource reaping (with the previous 1.5
minutes being log collection).

[1]: https://kubernetes.io/docs/concepts/workloads/pods/pod/#termination-of-pods
[2]: https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_installer/1255/pull-ci-openshift-installer-master-e2e-aws/3707/build-log.txt
[3]: https://gcsweb-ci.svc.ci.openshift.org/gcs/origin-ci-test/pr-logs/pull/openshift_installer/1255/pull-ci-openshift-installer-master-e2e-aws/3707/artifacts/
[4]: https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_installer/1045/pull-ci-openshift-installer-master-e2e-aws/3706/build-log.txt
[5]: https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_installer/1045/pull-ci-openshift-installer-master-e2e-aws/3706/artifacts/e2e-aws/installer/.openshift_install.log